Move coverage reporting to end of e2e testing
